+++ editor/src/point.cpp
@@ -49,58 +49,15 @@ Solarus::Point to_solarus_point(const QPoint& qt_point) {
  */
 QPoint round_8(const QPoint& point) {
 
-  // The division operator of QPoint already implements rounding to the
-  // nearest integer.
-  return point / 8 * 8;
+  // Make sure negative values don't get biased towards zero.
+  return floor_8(point + QPoint(4, 4));
 }
 
 /**
  * @overload
  */
 QPoint round_8(const QPointF& point) {
-
-  return round_8(point.toPoint());
-}
-
-/**
- * @brief Floors the coordinates of a point to the lower multiples of 8.
- * @param point A point.
- * @return The resulting point.
- */
-QPoint floor_8(const QPoint& point) {
-
-  return floor(point, QSize(8, 8));
-}
-
-/**
- * @overload
- */
-QPoint floor_8(const QPointF& point) {
-
-  return floor(point, QSize(8, 8));
-}
-
-/**
- * @brief Floors the coordinates of a point to the lower multiples of a size.
- * @param point A point.
- * @param size A size.
- * @return The resulting point.
- */
-QPoint floor(const QPoint& point, const QSize& size) {
-
-  return QPoint(point.x() - point.x() % size.width(),
-                point.y() - point.y() % size.height());
-}
-
-/**
- * @overload
- */
-QPoint floor(const QPointF& point, const QSize& size) {
-
-  return floor(
-      QPoint(qFloor(point.x()), qFloor(point.y())),
-      size
-  );
+    return round_8(point.toPoint());
 }
 
 /**
@@ -146,17 +103,17 @@ QPoint ceil(const QPointF& point, const QSize& size) {
 }
 
 /**
- * @brief Rounds the coordinates of a point down to the nearest lower multiple
+ * @brief Floors the coordinates of a point down to the nearest lower multiple
  * of given coordinates.
  *
- * round_down(QPoint(15, -2), 8, 8) is QPoint(8, -8).
+ * floor(QPoint(15, -2), QSize{8, 8) is QPoint(8, -8).
  *
  * @param point A point.
  * @param step_x Value where to round the x coordinate.
  * @param step_y Value where to round the y coordinate.
  * @return The resulting point.
  */
-QPoint round_down(const QPoint& point, const QSize& size) {
+QPoint floor(const QPoint& point, const QSize& size) {
 
   int mod_x = point.x() % size.width();
   int mod_y = point.y() % size.height();
@@ -176,11 +133,27 @@ QPoint round_down(const QPoint& point, const QSize& size) {
 /**
  * @overload
  */
-QPoint round_down(const QPointF& point, const QSize& size) {
+QPoint floor(const QPointF& point, const QSize& size) {
 
-  return round_down(point.toPoint(), size);
+  return floor(point.toPoint(), size);
 }
 
+/**
+ * @brief Floors the coordinates of a point to the lower multiples of 8.
+ * @param point A point.
+ * @return The resulting point.
+ */
+QPoint floor_8(const QPoint& point) {
+  return floor(point, QSize(8, 8));
 }
 
+/**
+ * @overload
+ */
+QPoint floor_8(const QPointF& point) {
+  return floor(point, QSize(8, 8));
 }
+
+}  // namespace Point
+
+}  // namespace SolarusEditor