+++ editor/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Solarus Quest Editor 2.0.1 (in progress)
+
+* Fix empty tileset selector (#1724).
+* Fix tiles position following the mouse when picked from the tileset view.
+* Fix wrong entity selection below entities just added (#1905).
+* Adjust combobox dropdown size when expanding folders.
+
 ## Solarus Quest Editor 2.0.0 (2025-04-01)
 
 ### New features